Muitos problemas envolvendo a criação de text box corrigidod. Diminui o tamanho do componente de edição separando as funções em arquivos separados
