+++ src/components/PageContent/components/PDFEdit/components/TextAreaComponent/TextArea.tsx
@@ -2,19 +2,21 @@ import React, { useEffect, useRef, useState } from 'react';
 import Draggable from 'react-draggable';
 
 interface TextAreaComponentProps {
-    x: number;
-    y: number;
+    x?: number;
+    y?: number;
     zoomLevel: number;
     onTextSubmit: (text: string) => void;
     fontsize?: number;
     fontColor?: string;
     onclick?: () => void;
     editableText?: string;
+    drawingRef?: any;
 }
 
-export const TextAreaComponent: React.FC<TextAreaComponentProps> = ({ x, y, zoomLevel, onTextSubmit, fontsize, fontColor, onclick, editableText}) => {
-    const [isFocused, setIsFocused] = useState(true); 
+export const TextAreaComponent: React.FC<TextAreaComponentProps> = ({ drawingRef, x, y, zoomLevel, onTextSubmit, fontsize, fontColor, onclick, editableText }) => {
+    const [isFocused, setIsFocused] = useState(true);
     const textAreaRef = useRef<HTMLTextAreaElement>(null);
+    const [position, setPosition] = useState({ x: x, y: y });
 
     const handleBlur = () => {
         setIsFocused(false);
@@ -46,13 +48,19 @@ export const TextAreaComponent: React.FC<TextAreaComponentProps> = ({ x, y, zoom
 
     useEffect(() => {
         if (textAreaRef.current && isFocused) {
-            textAreaRef.current.value = editableText || ""; 
+            textAreaRef.current.value = editableText || "";
             textAreaRef.current.focus();
         }
     }, [editableText, isFocused]);
 
     return (
-        <Draggable defaultPosition={{ x: x * zoomLevel, y: y * zoomLevel }}>
+        <Draggable 
+        defaultClassName='absolute'
+        position={{
+            x: (position.x!) * zoomLevel || 0,
+            y: (position.y!) * zoomLevel || 0
+        }}
+        onDrag={(e, data) => setPosition({ x: data.x, y: data.y })}>
             <textarea
                 onClick={onclick}
                 style={{
@@ -60,7 +68,7 @@ export const TextAreaComponent: React.FC<TextAreaComponentProps> = ({ x, y, zoom
                     fontSize: fontsize + 'px',
                     border: isFocused ? '2px dashed gray' : 'none',
                 }}
-                className='absolute overflow-hidden z-20 bg-transparent outline-none resize-none rounded-md'
+                className='relative h-fit w-fit overflow-hidden z-20 bg-transparent outline-none resize-none rounded-md'
                 ref={textAreaRef}
                 onFocus={handleFocus}
                 onBlur={handleBlur}
