+++ src/functions/index.ts
@@ -1,0 +1,9 @@
+export { handleRedo } from './handleRedo'
+export { renderDrawings } from './renderDrawings'
+export { renderPdf } from './renderPdf'
+export { distanceToLine } from './distanceLine'
+export { getTextClicked } from './getTextClicked'
+export { drawColor } from './drawColor'
+export { writeChanges } from './writeChanges'
+export { save } from './savePdf'
+export { handleText } from './handleText'
