+++ src/style/global.css
@@ -24,30 +24,30 @@ div::-webkit-scrollbar {
     background: #F4F4F4;
 }
 
-/* Thumb (bolinha) do slider */
 input[type="range"]::-webkit-slider-thumb {
     -webkit-appearance: none;
     height: 16px;
     width: 16px;
     border-radius: 50%;
-    background: #0d6efd; /* Cor da bolinha para o modo claro */
+    background: #0d6efd;
     cursor: pointer;
     transition: background-color 0.3s ease;
 }
 
 input[type="range"] {
-    -webkit-appearance: none; /* Remove o estilo padrão */
+    appearance: none;
+    -webkit-appearance: none; 
     width: 70%;
-    background: transparent; /* Remove o fundo do range */
+    background: transparent;
 }
 
 input[type="range"]::-webkit-slider-runnable-track {
-    background: #ddd; /* Cor do track para o modo claro */
+    background: #ddd; 
     border-radius: 10px;
 }
 
 input[type="range"]::-webkit-slider-thumb:hover {
-    background: #0a58ca; /* Cor ao passar o mouse */
+    background: #0a58ca; 
     box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.3);
 }
 
@@ -56,14 +56,14 @@ input[type="range"]::-webkit-slider-thumb:hover {
     height: 16px;
     width: 16px;
     border-radius: 50%;
-    background: #8a8a8a; /* Cor da bolinha para o modo claro */
+    background: #8a8a8a; 
     cursor: pointer;
     transition: background-color 0.3s ease;
 }
 
 
 .dark input[type="range"]::-webkit-slider-thumb:hover {
-    background: #8a8a8a; /* Cor ao passar o mouse */
+    background: #8a8a8a; 
     box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.3);
 }
